Modified ship controls to be more sensible. Recolored asteroids. ALSO, IMPORTANT: All update(int) methods changed to update(float). At first the parameter was the number of milliseconds since the last frame, but now it is the number of seconds since the last frame as a float. This makes it easier to step things like physics functions independently of the FPS (e.g. to move at 1m/s, add 1*delta to the position every frame)

+++ GameEngine/src/com/javagameengine/scene/component/Camera.java
@@ -156,7 +156,7 @@ public class Camera extends Component
 	}
 	
 	@Override
-	public void onUpdate(int delta)
+	public void onUpdate(float delta)
 	{
 	}
